pyproject.toml
[project]
name = "improved-ddgan"
version = "0.1.0"
description = "Improved DDGAN"
dependencies = [
    "torch==2.2.2+cu121",
    "torchvision==0.17.2+cu121",
    "ftfy>=6.2.3",
    "regex>=2024.7.24",
    "tqdm>=4.64.1",
    "clip @ git+https://github.com/openai/CLIP.git",
    "pillow>=10.2.0",
    "matplotlib>=3.9.2",
    "tensorboard>=2.17.1",
    "tensorboardx>=2.6.2.2",
    "lmdb>=1.5.1",
    "scipy>=1.14.1",
    "ninja>=1.11.1.1",
    "einops>=0.8.0",
    "pywavelets>=1.7.0",
    "cython>=3.0.11",
    "wandb>=0.17.8",
    "huggingface-hub>=0.24.6",
    "pycocotools>=2.0.8",
    "omegaconf>=2.3.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/improved_ddgan"]

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"
